wip: status banner
